# Increase coordinator Lambda default memory to improve performance and stability