--- a/open-tofu/variables.tf
+++ b/open-tofu/variables.tf
@@ -3,3 +3,9 @@ variable "dnssec_enabled" {
   description = "Enable DNSSEC"
   default     = true
 }
+
+variable "coordinator_memory_size" {
+  type        = number
+  description = "MPIC Coordinator Lambda Function Memory"
+  default     = 256
+}