@birgelee 2025-03-31
Thanks for bringing this to our attention. We intend to address this. We appreciate you helping with the AWS tuning. We had also noticed some timeouts particularly occurring  on high remote perspective counts.