@birgelee 2025-04-01
Thanks for contributing this. I confirmed it passes all integration tests and fixed a bug we previously had with it sometimes failing integration tests.

I did take the liberty of doubling the memory to 512 as I feel 80% memory pressure is not good and this could be exceeded if even more perspectives were added.